Add directory 'Onboarding' int he Components folder, relocate onboarding components into it, update all import statements
